refactor: renamed snapshot to projection
